Use Server components for properties page.
