+++ components/Pagination.jsx
@@ -1,18 +1,18 @@
-const Pagination = ({ page, pageSize, totalItems, onPageChange }) => {
-  const totalPages = Math.ceil(totalItems / pageSize);
+'use client';
 
-  const handlePageChange = (newPage) => {
-    if (newPage >= 1 && newPage <= totalPages) {
-      onPageChange(newPage);
-    }
-  };
+// TODO: change the buttons to Link components and server render.
+
+import { useRouter } from 'next/navigation';
+const Pagination = ({ page, pageSize, totalItems }) => {
+  const router = useRouter();
+  const totalPages = Math.ceil(totalItems / pageSize);
 
   return (
     <section className='container mx-auto flex justify-center items-center my-8'>
       <button
         className='mr-2 px-2 py-1 border border-gray-300 rounded'
         disabled={page === 1}
-        onClick={() => handlePageChange(page - 1)}
+        onClick={() => router.push(`/properties?page=${page - 1}`)}
       >
         Previous
       </button>
@@ -22,7 +22,7 @@ const Pagination = ({ page, pageSize, totalItems, onPageChange }) => {
       <button
         className='ml-2 px-2 py-1 border border-gray-300 rounded'
         disabled={page === totalPages}
-        onClick={() => handlePageChange(page + 1)}
+        onClick={() => router.push(`/properties?page=${page + 1}`)}
       >
         Next
       </button>
