+++ components/Properties.jsx
@@ -1,47 +1,12 @@
-'use client';
-import { useState, useEffect } from 'react';
 import PropertyCard from '@/components/PropertyCard';
-import Spinner from '@/components/Spinner';
 import Pagination from '@/components/Pagination';
 
-const Properties = () => {
-  const [properties, setProperties] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [page, setPage] = useState(1);
-  const [pageSize, setPageSize] = useState(6);
-  const [totalItems, setTotalItems] = useState(0);
+// NOTE: This has been changed to a server component and now recieves props from
+// the page parent so no need to make a fetch request to an API route handler.
 
-  useEffect(() => {
-    const fetchProperties = async () => {
-      try {
-        const res = await fetch(
-          `/api/properties?page=${page}&pageSize=${pageSize}`
-        );
-
-        if (!res.ok) {
-          throw new Error('Failed to fetch data');
-        }
-
-        const data = await res.json();
-        setProperties(data.properties);
-        setTotalItems(data.total);
-      } catch (error) {
-        console.log(error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchProperties();
-  }, [page, pageSize]);
-
-  const handlePageChange = (newPage) => {
-    setPage(newPage);
-  };
-
-  return loading ? (
-    <Spinner />
-  ) : (
+const Properties = ({ properties, total, page, pageSize }) => {
+  // TODO: fallback to loader
+  return (
     <section className='px-4 py-6'>
       <div className='container-xl lg:container m-auto px-4 py-6'>
         {properties.length === 0 ? (
@@ -53,12 +18,7 @@ const Properties = () => {
             ))}
           </div>
         )}
-        <Pagination
-          page={page}
-          pageSize={pageSize}
-          totalItems={totalItems}
-          onPageChange={handlePageChange}
-        />
+        <Pagination page={page} pageSize={pageSize} totalItems={total} />
       </div>
     </section>
   );
